Rename `extension` crate to `extension_host`
