+++ crates/activity_indicator/Cargo.toml
@@ -16,7 +16,7 @@ doctest = false
 anyhow.workspace = true
 auto_update.workspace = true
 editor.workspace = true
-extension.workspace = true
+extension_host.workspace = true
 futures.workspace = true
 gpui.workspace = true
 language.workspace = true
