+++ crates/extension_host/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "extension"
+name = "extension_host"
 version = "0.1.0"
 edition = "2021"
 publish = false
@@ -9,7 +9,7 @@ license = "GPL-3.0-or-later"
 workspace = true
 
 [lib]
-path = "src/extension_store.rs"
+path = "src/extension_host.rs"
 doctest = false
 
 [features]
