+++ crates/extension_host/build.rs
