+++ crates/extension_host/src/extension_indexed_docs_provider.rs
