+++ crates/extension_host/src/extension_manifest.rs
