+++ crates/extension_host/src/extension_settings.rs
