+++ crates/extension_host/src/extension_slash_command.rs
