+++ crates/extension_host/src/extension_store_test.rs
