+++ crates/extension_host/src/wasm_host/wit/since_v0_0_1.rs
