+++ crates/extension_host/src/wasm_host/wit/since_v0_1_0.rs
