+++ crates/extension_host/src/wasm_host/wit/since_v0_2_0.rs
